added neighbor sharing

+++ src/dgcp_handler.h
@@ -8,15 +8,15 @@
 pthread_t tid[2];
 pthread_mutex_t lock;
 
-extern void dgcp_send(int s, unsigned char ipv6[], int port, dgc_packet p2send);
+extern void dgcp_send(int s, unsigned char ipv6[], uint16_t port, dgc_packet p2send);
 extern void *dgcp_recv(void *arguments);
 
-extern void header_handler(int s, unsigned char ipv6[], int port, dgc_packet p2recv);
-extern void call_tlv_handler(int s, unsigned char ipv6[], int port, msg_body tlv, uint8_t type);
-extern void hello_handler(int s, unsigned char ipv6[], int port, msg_body tlv);
-extern void neighbor_handler(int s, unsigned char ipv6[], int port, msg_body tlv);
-extern void data_handler(int s, unsigned char ipv6[], int port, msg_body tlv);
-extern void ack_handler(int s, unsigned char ipv6[], int port, msg_body tlv);
-extern void goaway_handler(int s, unsigned char ipv6[], int port, msg_body tlv);
+extern void header_handler(int s, unsigned char ipv6[], uint16_t port, dgc_packet p2recv);
+extern void call_tlv_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv, uint8_t type);
+extern void hello_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv);
+extern void neighbor_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv);
+extern void data_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv);
+extern void ack_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv);
+extern void goaway_handler(int s, unsigned char ipv6[], uint16_t port, msg_body tlv);
 
 #endif
